SUPER COMMIT (please read the description)
--> cleanBoard()
Bug fix, remove the green cells when starts a new game.

--> insertNumber()
Bug fix, just decrement the value of missing cells, if the cell had no
value before.
Bug fix, if the cells are green and the number of cells missing are
different of 0, so remove the color green of the cells.

--> animateCell()
Delete code, of unuseless animations.

--> checkGameOver()
Add code, for highlight cells to green, when the game is over.

--> removeHighlightGreen() -  NEW FUNCTION
Used for highlight the cells to green.